fix: always bust cache when using a dpkg-packages file
If a dpkg-packages file is in use, the cache _must_ be busted. This is because the files in the repository are not guaranteed to be the same during each deploy - names can be reused - and the plugin does not know whether or not the deploy changes the files.

Instead of using a dpkg-packages file, users are encouraged to setup an apt repository and upload packages there.

Refs #23
Refs #32